make not specific to httpbin, divide code up
